Struct editor bugfixes, Struct/Field Def refactoring:
- StructEditorWidget: when an entry is passed to a dialog, or recieved from a dialog, a copy should be made so that it is not deleted when the dialog is?
- StructEditorWidget: when there is a field that is not padding and the label column was double clicked, it should instead edit the label of that row using the table view and not a watch entry dialog
- StructEditorWidget: now creates copy of the dialog entry to prevent unexpected deletion
- StructDetailModel: now uses the new method names for setting and getting the size of a field
- StructDetailModel: in remove padding fields, fields should only be removed if there is more than 0 fields to remove.
- StructDetailModel: indices have no internal pointer, so get field by row is necessary.
- FieldDef: changed m_structOffset to a u32 since it should never exceed the size of the ram
- FieldDef: When copying another field def, the size should be set to the same value - fixed.
- FieldDef: The length of the field should be stored for proper loading in the struct editor
- FieldDef: When loading a field def from json, if there is no "entry" it should m_entry should become null. Also, need to read the json into the m_entry member variable - fixed.
- StructDef: renamed variables for clarity

+++ Source/Structs/FieldDef.h
@@ -25,8 +25,8 @@ public:
   FieldDef& operator=(const FieldDef&) = delete;
   FieldDef& operator=(FieldDef&&) = delete;
 
-  size_t getOffset() const;
-  void setOffset(size_t offset);
+  u32 getOffset() const;
+  void setOffset(u32 offset);
   MemWatchEntry* getEntry() const;
   void setEntry(MemWatchEntry* entry);
   size_t getFieldSize() const;
@@ -40,7 +40,7 @@ public:
   void writeToJson(QJsonObject& json);
 
 private:
-  size_t m_structOffset;
+  u32 m_structOffset;
   size_t m_size;
   MemWatchEntry* m_entry;
 };
